feat: render markdown assets in terminal
