+++ src/render/code_block.rs
@@ -1,0 +1,260 @@
+use super::style::{
+    CODE_BLOCK_BG, CODE_BLOCK_FRAME_STYLE, CODE_COMMENT_STYLE, CODE_FUNCTION_STYLE,
+    CODE_KEYWORD_STYLE, CODE_NUMBER_STYLE, CODE_STRING_STYLE, CODE_TOKEN_RESET, PRIMARY_STYLE,
+    RESET,
+};
+
+/// 渲染普通代码块。
+///
+/// 参数:
+/// - `lang`: Markdown 代码块语言标识
+/// - `lines`: 代码块内容行
+///
+/// 返回:
+/// - 带终端样式的代码块文本
+pub(crate) fn render_code_block(lang: &str, lines: &[String]) -> String {
+    let label = if lang.is_empty() {
+        "code".to_string()
+    } else {
+        format!("code {lang}")
+    };
+    let header = format!("-- {label}");
+    let footer = "--";
+    let width = lines
+        .iter()
+        .map(|line| line.chars().count())
+        .chain([header.chars().count(), footer.chars().count()])
+        .max()
+        .unwrap_or(footer.len())
+        .max(24);
+    let mut output = String::new();
+    output.push_str(&render_code_block_frame(&header, width));
+    output.push('\n');
+    for line in lines {
+        output.push_str(&render_code_block_line_with_width(lang, line, width));
+        output.push('\n');
+    }
+    output.push_str(&render_code_block_frame(footer, width));
+    output.push('\n');
+    output
+}
+
+/// 渲染代码块边框行。
+///
+/// 参数:
+/// - `text`: 边框标签
+/// - `width`: 目标宽度
+///
+/// 返回:
+/// - 带样式的边框文本
+fn render_code_block_frame(text: &str, width: usize) -> String {
+    if text == "--" {
+        return format!("{CODE_BLOCK_FRAME_STYLE}{}{RESET}", "-".repeat(width));
+    }
+    let prefix = format!("{text} ");
+    format!(
+        "{CODE_BLOCK_FRAME_STYLE}{prefix}{}{RESET}",
+        "-".repeat(width.saturating_sub(prefix.chars().count()))
+    )
+}
+
+/// 按固定宽度渲染代码块内容行。
+///
+/// 参数:
+/// - `lang`: 语言标识
+/// - `line`: 原始代码行
+/// - `width`: 目标宽度
+///
+/// 返回:
+/// - 已补齐宽度并高亮的代码行
+fn render_code_block_line_with_width(lang: &str, line: &str, width: usize) -> String {
+    let line_width = line.chars().count();
+    let padding = " ".repeat(width.saturating_sub(line_width));
+    let highlighted = highlight_code_line(lang, line);
+    if highlighted.is_empty() {
+        format!("{CODE_BLOCK_BG}{}{RESET}", " ".repeat(width.max(1)))
+    } else {
+        format!("{CODE_BLOCK_BG}{highlighted}{padding}{RESET}")
+    }
+}
+
+/// 对单行代码做轻量语法高亮。
+///
+/// 参数:
+/// - `lang`: 语言标识
+/// - `line`: 代码行
+///
+/// 返回:
+/// - 带 ANSI 样式的代码行
+fn highlight_code_line(lang: &str, line: &str) -> String {
+    let lang = lang.trim().to_ascii_lowercase();
+    if lang.is_empty() {
+        return line.to_string();
+    }
+    let comment_marker = match lang.as_str() {
+        "py" | "python" | "sh" | "bash" | "zsh" | "fish" | "toml" | "yaml" | "yml" => Some('#'),
+        "rs" | "rust" | "js" | "ts" | "tsx" | "jsx" | "c" | "cpp" | "java" | "go" => None,
+        _ => None,
+    };
+    let mut output = String::new();
+    let chars = line.chars().collect::<Vec<_>>();
+    let mut index = 0;
+    while index < chars.len() {
+        if let Some(marker) = comment_marker {
+            if chars[index] == marker {
+                output.push_str(CODE_COMMENT_STYLE);
+                output.extend(chars[index..].iter());
+                output.push_str(CODE_TOKEN_RESET);
+                return output;
+            }
+        }
+        if index + 1 < chars.len() && chars[index] == '/' && chars[index + 1] == '/' {
+            output.push_str(CODE_COMMENT_STYLE);
+            output.extend(chars[index..].iter());
+            output.push_str(CODE_TOKEN_RESET);
+            return output;
+        }
+        if chars[index] == '"'
+            || chars[index] == '\''
+            || (chars[index] == '`'
+                && matches!(lang.as_str(), "js" | "ts" | "tsx" | "jsx" | "sh" | "bash"))
+        {
+            let quote = chars[index];
+            let start = index;
+            index += 1;
+            let mut escaped = false;
+            while index < chars.len() {
+                if escaped {
+                    escaped = false;
+                } else if chars[index] == '\\' {
+                    escaped = true;
+                } else if chars[index] == quote {
+                    index += 1;
+                    break;
+                }
+                index += 1;
+            }
+            output.push_str(CODE_STRING_STYLE);
+            output.extend(chars[start..index].iter());
+            output.push_str(CODE_TOKEN_RESET);
+            continue;
+        }
+        if chars[index].is_ascii_digit() {
+            let start = index;
+            index += 1;
+            while index < chars.len()
+                && (chars[index].is_ascii_alphanumeric() || matches!(chars[index], '_' | '.'))
+            {
+                index += 1;
+            }
+            output.push_str(CODE_NUMBER_STYLE);
+            output.extend(chars[start..index].iter());
+            output.push_str(CODE_TOKEN_RESET);
+            continue;
+        }
+        if is_code_word_start(chars[index]) {
+            let start = index;
+            index += 1;
+            while index < chars.len() && is_code_word_char(chars[index]) {
+                index += 1;
+            }
+            let token = chars[start..index].iter().collect::<String>();
+            let style = if code_keywords(&lang).contains(&token.as_str()) {
+                Some(CODE_KEYWORD_STYLE)
+            } else if matches!(
+                token.as_str(),
+                "true" | "false" | "null" | "None" | "Some" | "Ok" | "Err"
+            ) {
+                Some(CODE_NUMBER_STYLE)
+            } else if next_non_space_is_open_paren(&chars, index) {
+                Some(CODE_FUNCTION_STYLE)
+            } else {
+                None
+            };
+            if let Some(style) = style {
+                output.push_str(style);
+                output.push_str(&token);
+                output.push_str(CODE_TOKEN_RESET);
+            } else {
+                output.push_str(PRIMARY_STYLE);
+                output.push_str(&token);
+                output.push_str(CODE_TOKEN_RESET);
+            }
+            continue;
+        }
+        output.push(chars[index]);
+        index += 1;
+    }
+    output
+}
+
+/// 返回指定语言的关键词列表。
+///
+/// 参数:
+/// - `lang`: 语言标识
+///
+/// 返回:
+/// - 静态关键词数组
+fn code_keywords(lang: &str) -> &'static [&'static str] {
+    match lang {
+        "rs" | "rust" => &[
+            "as", "async", "await", "break", "const", "continue", "crate", "else", "enum", "fn",
+            "for", "if", "impl", "in", "let", "loop", "match", "mod", "move", "mut", "pub", "ref",
+            "return", "self", "Self", "static", "struct", "trait", "type", "unsafe", "use",
+            "where", "while",
+        ],
+        "py" | "python" => &[
+            "and", "as", "async", "await", "break", "class", "continue", "def", "elif", "else",
+            "except", "finally", "for", "from", "if", "import", "in", "is", "lambda", "not", "or",
+            "pass", "raise", "return", "try", "while", "with", "yield",
+        ],
+        "js" | "ts" | "tsx" | "jsx" => &[
+            "async", "await", "break", "case", "catch", "class", "const", "continue", "default",
+            "else", "export", "extends", "finally", "for", "from", "function", "if", "import",
+            "let", "new", "return", "switch", "throw", "try", "typeof", "var", "while",
+        ],
+        "sh" | "bash" | "zsh" | "fish" => &[
+            "case", "do", "done", "elif", "else", "esac", "fi", "for", "function", "if", "in",
+            "then", "while",
+        ],
+        "json" | "toml" | "yaml" | "yml" => &["true", "false", "null"],
+        _ => &[],
+    }
+}
+
+/// 判断字符是否可作为代码标识符起始。
+///
+/// 参数:
+/// - `ch`: 待判断字符
+///
+/// 返回:
+/// - 是否为标识符起始字符
+fn is_code_word_start(ch: char) -> bool {
+    ch.is_ascii_alphabetic() || ch == '_'
+}
+
+/// 判断字符是否可作为代码标识符组成部分。
+///
+/// 参数:
+/// - `ch`: 待判断字符
+///
+/// 返回:
+/// - 是否为标识符组成字符
+fn is_code_word_char(ch: char) -> bool {
+    ch.is_ascii_alphanumeric() || ch == '_'
+}
+
+/// 判断下一个非空白字符是否为左括号。
+///
+/// 参数:
+/// - `chars`: 字符数组
+/// - `index`: 起始索引
+///
+/// 返回:
+/// - 下一个非空白字符是否为 `(`
+fn next_non_space_is_open_paren(chars: &[char], mut index: usize) -> bool {
+    while index < chars.len() && chars[index].is_whitespace() {
+        index += 1;
+    }
+    chars.get(index) == Some(&'(')
+}
