+++ src/render/style.rs
@@ -1,0 +1,23 @@
+pub(crate) const RESET: &str = "\x1b[0m";
+pub(crate) const PRIMARY_STYLE: &str = "\x1b[38;5;189m";
+pub(crate) const SECONDARY_STYLE: &str = "\x1b[36m";
+pub(crate) const TERTIARY_STYLE: &str = "\x1b[35m";
+pub(crate) const HEADER_STYLE: &str = "\x1b[1m\x1b[35m";
+pub(crate) const INLINE_CODE_STYLE: &str = SECONDARY_STYLE;
+pub(crate) const LINK_LABEL_STYLE: &str = "\x1b[4m\x1b[38;5;81m";
+pub(crate) const URL_STYLE: &str = "\x1b[2m\x1b[38;5;75m";
+pub(crate) const IMAGE_STYLE: &str = "\x1b[1m\x1b[38;5;213m";
+pub(crate) const MATH_STYLE: &str = "\x1b[38;5;117m";
+pub(crate) const BOLD_STYLE: &str = "\x1b[1m\x1b[34m";
+pub(crate) const ITALIC_STYLE: &str = "\x1b[3m\x1b[38;5;250m";
+pub(crate) const STRIKE_STYLE: &str = "\x1b[9m";
+pub(crate) const CODE_BLOCK_BG: &str = "";
+pub(crate) const CODE_BLOCK_FRAME_STYLE: &str = SECONDARY_STYLE;
+pub(crate) const CODE_TOKEN_RESET: &str = "\x1b[0m";
+pub(crate) const CODE_KEYWORD_STYLE: &str = "\x1b[38;2;196;167;231m";
+pub(crate) const CODE_FUNCTION_STYLE: &str = "\x1b[38;2;156;207;216m";
+pub(crate) const CODE_STRING_STYLE: &str = "\x1b[38;2;166;214;160m";
+pub(crate) const CODE_NUMBER_STYLE: &str = "\x1b[38;2;246;193;119m";
+pub(crate) const CODE_COMMENT_STYLE: &str = "\x1b[2m\x1b[38;2;110;106;134m";
+pub(crate) const TABLE_BORDER_STYLE: &str = "\x1b[2m";
+pub(crate) const ASSET_ERROR_STYLE: &str = "\x1b[31m";
